Lab 1 - New Header to the Main Page
I chose to add a list listing 5 reasons as to why I love CS 193, as well as adding a header that contained my name and course. I also decided to keep the placeholder information in case I need it later on. I will remove it later if I need to.